% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_complexity_linter.R
\name{print.table_complexity_report}
\alias{print.table_complexity_report}
\title{Print a table complexity report}
\usage{
\method{print}{table_complexity_report}(x, ...)
}
\arguments{
\item{x}{A \code{table_complexity_report} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a table complexity report
}
