% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_complexity_linter.R
\docType{package}
\name{XLintR-package}
\alias{XLintR}
\alias{XLintR-package}
\title{Excel table complexity linter}
\description{
Functions for detecting table-like regions in Excel workbooks and
assessing their structural complexity.
}
\author{
\strong{Maintainer}: NISRA TechLab \email{techlab@nisra.gov.uk}

Authors:
\itemize{
  \item NISRA TechLab \email{techlab@nisra.gov.uk}
}

}
\keyword{internal}
